+++ tests/header/public_header_test.cpp
@@ -1,0 +1,10 @@
+// SPDX-FileCopyrightText: 2026 Alexandr Savca
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+#ifndef LIBPKGSTATE_SOURCE_TEST_HEADER
+#error "LIBPKGSTATE_SOURCE_TEST_HEADER is required"
+#endif
+
+#include LIBPKGSTATE_SOURCE_TEST_HEADER
+
+int main() {}